requirements.txt
supabase
langchain
langchain-openai
pypdf
python-dotenv
tiktoken
python-docx
langchain-community